rename and add pagination
Issue: AAH-2509
